adding ReAct agent
