+++ libs/superagent/.env.example
@@ -5,6 +5,7 @@ DATABASE_MIGRATION_URL=
 JWT_SECRET=
 # Mandatory for running Open Source LLMs
 OPENROUTER_API_KEY=
+HUGGINGFACE_API_KEY=
 # Mandatory for Neon DB
 DATABASE_SHADOW_URL= 
 # Memory (mandatory)
