+++ libs/superagent/app/agents/base.py
@@ -1,13 +1,51 @@
+import datetime
+import json
+from abc import ABC, abstractmethod
 from typing import Any, List
 
+from decouple import config
+from langchain.memory.motorhead_memory import MotorheadMemory
+from langchain.schema import SystemMessage
+from slugify import slugify
+
+from app.datasource.types import VALID_UNSTRUCTURED_DATA_TYPES
+from app.models.tools import DatasourceInput
+from app.tools import (
+    TOOL_TYPE_MAPPING,
+    create_pydantic_model_from_object,
+    create_tool,
+)
+from app.tools.datasource import DatasourceTool, StructuredDatasourceTool
 from app.utils.prisma import prisma
-from app.utils.streaming import CustomAsyncIteratorCallbackHandler
 from prisma.models import Agent, AgentDatasource, AgentLLM, AgentTool
 
 DEFAULT_PROMPT = (
     "You are a helpful AI Assistant, anwer the users questions to "
     "the best of your ability."
 )
+
+
+def recursive_json_loads(data):
+    if isinstance(data, str):
+        try:
+            data = json.loads(data)
+        except json.JSONDecodeError:
+            return data
+    if isinstance(data, dict):
+        return {k: recursive_json_loads(v) for k, v in data.items()}
+    if isinstance(data, list):
+        return [recursive_json_loads(v) for v in data]
+    return data
+
+
+class AbstractAgent(ABC):
+    @abstractmethod
+    async def _get_llm(self, agent_llm: AgentLLM, model: str) -> Any:
+        pass
+
+    @abstractmethod
+    async def get_agent(self):
+        pass
 
 
 class AgentBase:
@@ -17,27 +55,115 @@ class AgentBase:
         session_id: str = None,
         enable_streaming: bool = False,
         output_schema: str = None,
-        callback: CustomAsyncIteratorCallbackHandler = None,
     ):
         self.agent_id = agent_id
         self.session_id = session_id
         self.enable_streaming = enable_streaming
         self.output_schema = output_schema
-        self.callback = callback
 
     async def _get_tools(
-        self, agent_datasources: List[AgentDatasource], agent_tools: List[AgentTool]
+        self,
+        agent_datasources: List[AgentDatasource],
+        agent_tools: List[AgentTool],
     ) -> List:
-        raise NotImplementedError
+        tools = []
+        for agent_datasource in agent_datasources:
+            tool_type = (
+                DatasourceTool
+                if agent_datasource.datasource.type in VALID_UNSTRUCTURED_DATA_TYPES
+                else StructuredDatasourceTool
+            )
 
-    async def _get_llm(self, agent_llm: AgentLLM, model: str) -> Any:
-        raise NotImplementedError
+            metadata = (
+                {
+                    "datasource_id": agent_datasource.datasource.id,
+                    "options": agent_datasource.datasource.vectorDb.options
+                    if agent_datasource.datasource.vectorDb
+                    else {},
+                    "provider": agent_datasource.datasource.vectorDb.provider
+                    if agent_datasource.datasource.vectorDb
+                    else None,
+                    "query_type": "document",
+                }
+                if tool_type == DatasourceTool
+                else {"datasource": agent_datasource.datasource}
+            )
+            tool = tool_type(
+                metadata=metadata,
+                args_schema=DatasourceInput,
+                name=slugify(agent_datasource.datasource.name),
+                description=agent_datasource.datasource.description,
+                return_direct=False,
+            )
+            tools.append(tool)
+        for agent_tool in agent_tools:
+            tool_info = TOOL_TYPE_MAPPING.get(agent_tool.tool.type)
+            if agent_tool.tool.type == "FUNCTION":
+                metadata = recursive_json_loads(agent_tool.tool.metadata)
+                args = metadata.get("args", {})
+                PydanticModel = create_pydantic_model_from_object(args)
+                tool = create_tool(
+                    tool_class=tool_info["class"],
+                    name=metadata.get("functionName"),
+                    description=agent_tool.tool.description,
+                    metadata=agent_tool.tool.metadata,
+                    args_schema=PydanticModel,
+                    return_direct=agent_tool.tool.returnDirect,
+                )
+            else:
+                tool = create_tool(
+                    tool_class=tool_info["class"],
+                    name=slugify(agent_tool.tool.name),
+                    description=agent_tool.tool.description,
+                    metadata=agent_tool.tool.metadata,
+                    args_schema=tool_info["schema"],
+                    session_id=f"{self.agent_id}-{self.session_id}"
+                    if self.session_id
+                    else f"{self.agent_id}",
+                    return_direct=agent_tool.tool.returnDirect,
+                )
+            tools.append(tool)
+        return tools
 
     async def _get_prompt(self, agent: Agent) -> str:
-        raise NotImplementedError
+        if self.output_schema:
+            if agent.prompt:
+                content = (
+                    f"{agent.prompt}\n\n"
+                    "Always answer using the below output schema. "
+                    "No other characters allowed.\n\n"
+                    "Here is the output schema:\n"
+                    f"{self.output_schema}"
+                    "\n\nCurrent date: "
+                    f"{datetime.datetime.now().strftime('%Y-%m-%d')}"
+                )
+            else:
+                content = (
+                    f"{DEFAULT_PROMPT}\n\n"
+                    "Always answer using the below output schema. "
+                    "No other characters allowed.\n\n"
+                    "Here is the output schema:\n"
+                    f"{self.output_schema}"
+                    "\n\nCurrent date: "
+                    f"{datetime.datetime.now().strftime('%Y-%m-%d')}"
+                )
+        else:
+            content = agent.prompt or DEFAULT_PROMPT
+            content = f"{content}" f"\n\n{datetime.datetime.now().strftime('%Y-%m-%d')}"
+        return SystemMessage(content=content)
 
     async def _get_memory(self) -> List:
-        raise NotImplementedError
+        memory = MotorheadMemory(
+            session_id=f"{self.agent_id}-{self.session_id}"
+            if self.session_id
+            else f"{self.agent_id}",
+            memory_key="chat_history",
+            url=config("MEMORY_API_URL"),
+            return_messages=True,
+            output_key="output",
+        )
+        await memory.init()
+        return memory
 
     async def get_agent(self):
         agent_config = await prisma.agent.find_unique_or_raise(
@@ -52,24 +178,26 @@ class AgentBase:
         )
 
         if agent_config.llms[0].llm.provider in ["OPENAI", "AZURE_OPENAI"]:
-            from app.agents.langchain import LangchainAgent
+            from app.agents.function_calling_agent import FunctionCallingAgent
 
-            agent = LangchainAgent(
+            agent = FunctionCallingAgent(
                 agent_id=self.agent_id,
                 session_id=self.session_id,
                 enable_streaming=self.enable_streaming,
                 output_schema=self.output_schema,
-                callback=self.callback,
             )
         else:
-            from app.agents.superagent import SuperagentAgent
+            # this part is for Open Source LLMs
+            from app.agents.react_agent import ReActAgent
 
-            agent = SuperagentAgent(
+            agent = ReActAgent(
                 agent_id=self.agent_id,
                 session_id=self.session_id,
                 enable_streaming=self.enable_streaming,
                 output_schema=self.output_schema,
-                callback=self.callback,
             )
 
         return await agent.get_agent(config=agent_config)
+
+    async def _get_llm(self):
+        NotImplementedError
